apps/importer/requirements.txt
progressbar2
pymongo